impr: chat messages ui tweak
