+++ apps/masterbots.ai/components/chat-list.tsx
@@ -1,9 +1,9 @@
 import { type Message } from 'ai'
 
 import { ChatMessage } from '@/components/chat-message'
+import { createMessagePairs } from '@/lib/utils'
 import { Chatbot } from 'mb-genql'
 import React from 'react'
-import { createMessagePairs } from '@/lib/utils'
 import { ChatAccordion } from './chat-accordion'
 import { ShortMessage } from './short-message'
 
@@ -39,13 +39,16 @@ export function ChatList({
     <div className="relative max-w-3xl px-4 mx-auto">
       {pairs.map((pair: MessagePair, key: number) => (
         <div key={key}>
-          <ChatAccordion defaultState className="border-none mb-4">
+          <ChatAccordion defaultState={key === 0} className="border-none mb-4">
+            {/* Thread Title */}
             <ChatMessage
               actionRequired={false}
               chatbot={chatbot}
               message={pair.userMessage}
               sendMessageFromResponse={sendMessageFromResponse}
             />
+
+            {/* Thread Description */}
             <div className="opacity-50 overflow-hidden text-sm">
               {pair.chatGptMessage[0]?.content ? (
                 <div className="flex-1 px-1 ml-4 space-y-2 overflow-hidden text-left">
@@ -55,16 +58,18 @@ export function ChatList({
                 ''
               )}
             </div>
+
+            {/* Thread Content */}
             <div className="max-h-[75vh] scrollbar">
               {pair.chatGptMessage.length > 0
                 ? pair.chatGptMessage.map((message, index) => (
-                    <ChatMessage
-                      key={index}
-                      chatbot={chatbot}
-                      message={message}
-                      sendMessageFromResponse={sendMessageFromResponse}
-                    />
-                  ))
+                  <ChatMessage
+                    key={index}
+                    chatbot={chatbot}
+                    message={message}
+                    sendMessageFromResponse={sendMessageFromResponse}
+                  />
+                ))
                 : ''}
             </div>
           </ChatAccordion>
